move sortParticipants method to composable
